#788 force proxy BEFORE calling Navigator.appendBasicAuthIfNeeded()
