+++ src/test/java/com/codeborne/selenide/WebDriverRunnerTest.java
@@ -16,6 +16,7 @@ import org.openqa.selenium.support.events.WebDriverEventListener;
 
 import java.net.URL;
 
+import static com.codeborne.selenide.Configuration.FileDownloadMode.HTTPGET;
 import static com.codeborne.selenide.Selenide.open;
 import static com.codeborne.selenide.WebDriverRunner.FIREFOX;
 import static com.codeborne.selenide.WebDriverRunner.HTMLUNIT;
@@ -42,6 +43,8 @@ class WebDriverRunnerTest implements WithAssertions {
 
     WebDriverRunner.webdriverContainer = spy(new WebDriverThreadLocalContainer());
     doReturn(null).when((JavascriptExecutor) driver).executeScript(anyString(), any());
+    Configuration.proxyEnabled = false;
+    Configuration.fileDownload = HTTPGET;
   }
 
   @AfterEach
