+++ src/test/java/integration/BasicAuthTest.java
@@ -5,7 +5,6 @@ import com.codeborne.selenide.Credentials;
 import org.junit.jupiter.api.Test;
 
 import static com.codeborne.selenide.Condition.text;
-import static com.codeborne.selenide.Configuration.FileDownloadMode.PROXY;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.open;
 
@@ -18,21 +17,21 @@ class BasicAuthTest extends IntegrationTest {
 
   @Test
   void canPassBasicAuth_via_proxy() {
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
     open("/basic-auth/hello", "", "scott", "tiger");
     $("body").shouldHave(text("Hello, scott:tiger!"));
   }
 
   @Test
   void canAuthUsingProxyWithLoginAndPassword() {
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
     open("/basic-auth/hello", AuthenticationType.BASIC, "scott", "tiger");
     $("body").shouldHave(text("Hello, scott:tiger!"));
   }
 
   @Test
   void canAuthUsingProxyWithCredentials() {
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
     Credentials credentials = new Credentials("scott", "tiger");
     open("/basic-auth/hello", AuthenticationType.BASIC, credentials);
     $("body").shouldHave(text("Hello, scott:tiger!"));
@@ -42,7 +41,7 @@ class BasicAuthTest extends IntegrationTest {
 
   @Test
   void canSwitchToAnotherBasicAuth() {
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
     open("/basic-auth/hello", AuthenticationType.BASIC, new Credentials("scott", "tiger"));
     $("body").shouldHave(text("Hello, scott:tiger!"));
     open("/basic-auth/hello2", AuthenticationType.BASIC, new Credentials("scott2", "tiger2"));
@@ -51,7 +50,7 @@ class BasicAuthTest extends IntegrationTest {
 
   @Test
   void removesPreviousBasicAuthHeaders() {
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
     open("/basic-auth/hello", AuthenticationType.BASIC, new Credentials("scott", "tiger"));
     $("body").shouldHave(text("Hello, scott:tiger!"));
     open("/headers/hello3");
