+++ src/test/java/integration/FileDownloadViaProxyTest.java
@@ -9,7 +9,6 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
-import static com.codeborne.selenide.Configuration.FileDownloadMode.PROXY;
 import static com.codeborne.selenide.Selectors.byText;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.open;
@@ -24,7 +23,7 @@ class FileDownloadViaProxyTest extends IntegrationTest {
   void setUp() {
     assumeFalse(isPhantomjs()); // Why it's not working? It's magic for me...
 
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
     openFile("page_with_uploads.html");
   }
