+++ src/test/java/integration/proxy/ProxyServerUsageTest.java
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.codeborne.selenide.Configuration.FileDownloadMode.PROXY;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.WebDriverRunner.closeWebDriver;
 import static com.codeborne.selenide.WebDriverRunner.getSelenideProxy;
@@ -26,7 +25,7 @@ class ProxyServerUsageTest extends IntegrationTest {
   void setUp() {
     assumeFalse(isPhantomjs()); // I don't know why, but PhantomJS seems to ignore proxy
     closeWebDriver();
-    switchToDownloadMode(PROXY);
+    toggleProxy(true);
   }
 
   @Test
